refactor: CsvMeasurementWriter only creates cluster file if configured (#927)

+++ Examples/Io/Csv/include/ActsExamples/Io/Csv/CsvMeasurementWriter.hpp
@@ -52,7 +52,7 @@ class CsvMeasurementWriter final : public WriterT<MeasurementContainer> {
     /// Where to place output files
     std::string outputDir;
     /// Number of decimal digits for floating point precision in output.
-    size_t outputPrecision = std::numeric_limits<float>::max_digits10;
+    int outputPrecision = std::numeric_limits<float>::max_digits10;
   };
 
   /// Constructor with
